requirements.txt
Django==3.2.4
djangorestframework==3.12.4
mysqlclient==2.0.3
drf-yasg==1.21.5
PyMySQL==1.0.3
PyYAML==6.0
urllib3==1.26.6
django-rest-swagger
